refactor(RVCV-128): yml aws 설정 변경
