+++ api-module/src/main/resources/application.yml
@@ -12,5 +12,16 @@ spring:
     access-token-validity-time: 1h
     refresh-token-validity-time: 14d
 
-  autoconfigure:
-    exclude: org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration
+autoconfigure:
+  exclude: org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration
+
+aws:
+  credentials:
+    accessKey: ${S3_ACCESS_KEY}
+    secretKey: ${S3_SECRET_ACCESS_KEY}
+  s3:
+    bucket: romantic-pipe-review-bucket
+  region:
+    static: ap-northeast-2
+  stack:
+    auto: false
